Move payment-page code to Pledge
