--- .netlify/state.json
@@ -1,3 +1,0 @@
-{
-	"siteId": "f3300d64-9b8a-4a8d-a5f1-6c7cdc0fe002"
-}
